pyproject.toml
[tool.poetry]
name = "ksnb"
version = "0.1.0"
description = ""
authors = ["Your Name Here <shinykim63@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.0"
django-debug-toolbar = "^4.0.0"
djangorestframework = "^3.14.0"
strawberry-graphql = {extras = ["debug-server"], version = "^0.177.1"}
strawberry-graphql-django = "^0.9.4"
django-cors-headers = "^3.14.0"
pyjwt = "^2.7.0"
requests = "^2.30.0"
django-environ = "^0.10.0"
dj-database-url = "^2.0.0"
psycopg2-binary = "^2.9.6"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
sentry-sdk = "^1.22.2"
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
